+++ application/controllers/register.php
@@ -55,7 +55,7 @@ class Register extends CI_Controller
     }
     elseif($_POST) { $this->data['register_errors'] = ($_POST)?$this->form_validation->error_array():'No data submitted.'; }
     // Load View
-    $this->data['title'] = "Login / Register";
+    $this->data['title'] = "";
     if(isset($_REQUEST['ajax']) && $_REQUEST['ajax'] == 1)
     {
       $this->load->view('includes/functions');